Extract uploadFilesToCloudinary to thunkCreators
